Adds usage of SimpleCocoaWindow in Vulkan renderer on macOS
This allows to initialise renderer and on M1 pass until fail
related to a couple of required extensions being not available.

One extension absence likely has workaround as it is deprecated
extension which has a newer interface in modern version of Vulkan,
another one requires MoltenVk implementation which is in limbo
so far because on iOS required API is private.